requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
python-dotenv
watchdog
toml
streamlit-autorefresh
